[Toolchain] Combine the requests to the toolchain into ToolchainProvider
This commit combines the requests to the toolchain into ToolchainProvider.

ONE-vscode-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>
